Value::GetNumberOfElements should use size_t and not unsigned long.
This was triggering an assert on Windows.

+++ lib/Interpreter/Value.cpp
@@ -174,16 +174,16 @@ namespace cling {
     return isValid() && Ctx.hasSameType(getType(), Ctx.VoidTy);
   }
 
-  unsigned long Value::GetNumberOfElements() const {
+  size_t Value::GetNumberOfElements() const {
     if (const clang::ConstantArrayType* ArrTy
         = llvm::dyn_cast<clang::ConstantArrayType>(getType())) {
-      llvm::APInt arrSize(sizeof(unsigned long)*8, 1);
+      llvm::APInt arrSize(sizeof(size_t)*8, 1);
       do {
         arrSize *= ArrTy->getSize();
         ArrTy = llvm::dyn_cast<clang::ConstantArrayType>(ArrTy->getElementType()
                                                          .getTypePtr());
       } while (ArrTy);
-      return (unsigned long)arrSize.getZExtValue();
+      return static_cast<size_t>(arrSize.getZExtValue());
     }
     return 1;
   }
